+++ LeetCode/719_找出第K小的数对距离.cpp
@@ -6,12 +6,12 @@ class Solution {
 public:
     int smallestDistancePair(vector<int>& nums, int k) {
         sort(nums.begin(), nums.end());
-        int n = nums.size(), left = 0, right = nums.back() - nums.front();
+        int n = static_cast<int>(nums.size()), left = 0, right = nums.back() - nums.front();
         while (left <= right) {
             int mid = (left + right) / 2;
             int cnt = 0;
             for (int j = 0; j < n; j++) {
-                int i = lower_bound(nums.begin(), nums.begin() + j, nums[j] - mid) - nums.begin();
+                int i = static_cast<int>(lower_bound(nums.begin(), nums.begin() + j, nums[j] - mid) - nums.begin());
                 cnt += j - i;
             }
             if (cnt >= k) {
